requirements.txt
pandas>=2.0.0
requests>=2.31.0
python-dotenv>=1.0.0
phonenumbers>=8.13.0
beautifulsoup4>=4.12.0
tldextract>=5.0.0
Levenshtein>=0.21.0
googlemaps>=4.10.0
apify-client>=1.6.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
